% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{asd_prevalence}
\alias{asd_prevalence}
\title{Observations Across Published Studies of Autism Spectrum Disorder}
\format{
A tibble with 56 rows and 25 variables:
\describe{
\item{year}{}
\item{authors}{}
\item{country}{}
\item{pop}{}
\item{lowerage,upperage,medianage}{}
\item{asdn}{}
\item{normaliq}{}
\item{gender}{}
\item{report_prev}{}
\item{95ci}{}
\item{l95,u95}{}
\item{icd_10}{}
\item{dsm_iii,dsm_iv,dsm_iv_tr,dsm_five}{}
\item{screener}{}
\item{study}{}
\item{crude_prev,true_prev}{}
\item{cum_mean}{}
\item{order}{}
}
}
\usage{
asd_prevalence
}
\description{
Observations Across Published Studies of Autism Spectrum Disorder
}
\examples{
# Here is a glimpse at the data
# available in `asd_prevalence`
dplyr::glimpse(asd_prevalence)
}
\keyword{datasets}
