% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{chopsticks}
\alias{chopsticks}
\title{Efficiency of using chopsticks by length}
\format{
A tibble with 186 rows and 3 variables:
\describe{
\item{individual}{}
\item{length_amt}{}
\item{efficiency}{}
}
}
\usage{
chopsticks
}
\description{
Efficiency of using chopsticks by length
}
\examples{
# Here is a glimpse at the data
# available in `chopsticks`
dplyr::glimpse(chopsticks)
}
\keyword{datasets}
