% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_excel_file.R
\name{create_excel_file}
\alias{create_excel_file}
\title{Write an example Excel file based on \code{us_cities}}
\usage{
create_excel_file(path = NULL)
}
\arguments{
\item{path}{The path to which the Excel file will be written. If no \code{path} is
provided then the file will be written to the working directory. The
resulting filename will be \code{us_cities.xlsx}.}
}
\description{
Write an example Excel file based on \code{us_cities}
}
