% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_to_csv.R
\name{dataset_to_csv}
\alias{dataset_to_csv}
\title{Write an edr dataset to a CSV file}
\usage{
dataset_to_csv(dataset, path = NULL)
}
\arguments{
\item{dataset}{The name of a dataset in the \pkg{edr} package.}

\item{path}{The path to which the dataset will be written. If no \code{path} is
provided then the dataset will be written to the working directory. The
resulting filename will be the name of the dataset with the \code{.csv} extension.}
}
\description{
Write an edr dataset to a CSV file
}
