% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsp_code_rmd.R
\name{dsp_code_rmd}
\alias{dsp_code_rmd}
\title{Write an R Markdown file containing code listings for a chapter}
\usage{
dsp_code_rmd(chapter = 1:12, path = NULL)
}
\arguments{
\item{path}{The path to which the R Markdown file will be written. If no
\code{path} is provided then the file will be written to the working directory.}
}
\description{
Write an R Markdown file containing code listings for a chapter
}
