% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dspatterns-package.R
\docType{package}
\name{dspatterns-package}
\alias{dspatterns}
\alias{dspatterns-package}
\title{dspatterns: A Data Science Pattern Language: Designing Projects with R}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dspatterns/dspatterns}
  \item Report bugs at \url{https://github.com/dspatterns/dspatterns/issues}
}

}
\author{
\strong{Maintainer}: Alison Hill \email{apresstats@gmail.com} (\href{https://orcid.org/0000-0002-8082-1890}{ORCID})

Authors:
\itemize{
  \item Richard Iannone \email{riannone@me.com} (\href{https://orcid.org/0000-0003-3925-190X}{ORCID})
}

}
\keyword{internal}
