% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr_regex_gadget.R
\name{edr_regexplain_gadget}
\alias{edr_regexplain_gadget}
\title{Use the RegExplain gadget with some example text}
\usage{
edr_regexplain_gadget(text = NULL)
}
\arguments{
\item{text}{Text to experiment with in the RegExplain gadget. If nothing is
provided then the following sentence/string will be used:
\code{"The quick brown fox jumps over the lazy dog"}.}
}
\description{
Use the RegExplain gadget with some example text
}
