% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{employment}
\alias{employment}
\title{US employment figures from 1941 to 2010}
\format{
A tibble with 71 rows and 6 variables:
\describe{
\item{year}{The year for which the employment values apply.}
\item{population}{The total population of employable citizens.}
\item{employed}{The amount of citizens employed during the year.}
\item{agriculture,nonagriculture}{The amount of citizens employed in the
agricultural sector and those that were not}
\item{unemployed}{The amount of employable citizens that were unemployed
during the year.}
}
}
\usage{
employment
}
\description{
A dataset originally derived from thew USA Bureau of Labor Statistics. Data
obtained in tabular form from \url{https://datahub.io/core/employment-us} but
only a subset of columns was used here.
}
\examples{
# Here is a glimpse at the data
# available in `employment`
dplyr::glimpse(employment)
}
\keyword{datasets}
