% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{german_cities}
\alias{german_cities}
\title{Populations of large German cities}
\format{
A tibble with 79 rows and 4 variables:
\describe{
\item{name,state}{The id value for the order; each order can take multiple
rows (where each row is for the sale of an individual item)}
\item{pop_2015,pop_2011}{The census populations of each city in 2015 and
in 2011}
}
}
\usage{
german_cities
}
\description{
This dataset contains a population data on a selection of the largest cities
in Germany. The name and state of each city are provided as factor columns.
Population values are taken from 2011 Census data and 2015 estimate data.
Data obtained from the Wikipedia page at
\url{https://en.wikipedia.org/wiki/List_of_cities_in_Germany_by_population}.
}
\examples{
# Here is a glimpse at the data
# available in `german_cities`
dplyr::glimpse(german_cities)
}
\keyword{datasets}
