% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{maze}
\alias{maze}
\title{The maze dataset}
\format{
A tibble with 381 rows and 12 variables:
\describe{
\item{Study.ID}{The identifier for the study.}
\item{CA}{CA}
\item{VIQ}{VIQ}
\item{DX}{DX}
\item{Activity}{Activity}
\item{Content}{Content}
\item{Filler}{Filler}
\item{REP}{REP}
\item{REV}{REV}
\item{FS}{FS}
\item{Cued, Not.Cued}{Cued, Not.Cued}
}
}
\usage{
maze
}
\description{
The maze dataset
}
\examples{
# Here is a glimpse at the data
# available in `maze`
dplyr::glimpse(maze)
}
\keyword{datasets}
