% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pitchfork}
\alias{pitchfork}
\title{Album reviews from the Pitchfork website (1999-2018)}
\format{
A tibble with 20,852 rows and 7 variables:
\describe{
\item{artist,album,year}{The album artist, album name, and its year of
release.}
\item{genre}{One or more comma-separated musical genres applied to the
album by the reviewer.}
\item{score}{The reviewer score for the album (from 0 to 10).}
\item{date}{The publication date of the review.}
\item{link}{A link to the album review on the Pitchfork website.}
}
}
\usage{
pitchfork
}
\description{
Album reviews from the Pitchfork website (1999-2018)
}
\examples{
# Here is a glimpse at the data
# available in `pitchfork`
dplyr::glimpse(pitchfork)
}
\keyword{datasets}
