% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Yearly total rainfall amounts for seven cities in Canada}
\format{
A tibble with 25 rows and 8 variables:
\describe{
\item{year}{The year for which the total rainfall amount applies.}
\item{r_vancouver,r_calgary,r_kenora,r_toronto,r_montreal,r_halifax,r_stjohns}{
Total rainfall amounts (in millimeters) for the cities Vancouver, BC;
Calgary, AB; Kenora, ON; Toronto, ON; Montreal, QC; Halifax, NS; and
St. John's, NL.}
}
}
\usage{
rainfall
}
\description{
Yearly total rainfall amounts for seven cities in Canada
}
\examples{
# Here is a glimpse at the data
# available in `rainfall`
dplyr::glimpse(rainfall)
}
\keyword{datasets}
