% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{resto_reviews}
\alias{resto_reviews}
\title{Text from restaurant reviews for Momofuku Noodle Bar in Toronto}
\format{
A vector of length 40.
}
\usage{
resto_reviews
}
\description{
This dataset is a character vector of 40 restaurant reviews for the Momufuku
Noodle Bar located in Toronto, Canada. Reviews were obtained from Yelp and
retrieved from \url{https://www.yelp.ca/biz/momofuku-noodle-bar-toronto}.
}
\keyword{datasets}
