% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sales}
\alias{sales}
\title{Item sales from a shop that doesn't actually exist}
\format{
A tibble with 13,306 rows and 6 variables:
\describe{
\item{order_id}{The id value for the order; each order can take multiple
rows (where each row is for the sale of an individual item)}
\item{date,time}{The date and time of the order}
\item{item_id}{The id value for the item sold}
\item{price}{The price of the item sold}
}
}
\usage{
sales
}
\description{
This contains synthetic sales data for the month of January in 2019. Each row
constitutes an individual item sold. Multiple items could be sold as part of
a single order; the date and time indicate the time of the order. The \code{price}
is the sell price for the item.
}
\examples{
# Here is a glimpse at the data
# available in `sales`
dplyr::glimpse(sales)
}
\keyword{datasets}
