% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_cities}
\alias{us_cities}
\title{Larger US cities/towns and their populations}
\format{
A tibble with 765 rows and 5 variables:
\describe{
\item{city}{The name of the city, town, or unincorporated population
center}
\item{state_id,state_name}{The state or territory's USPS postal
abbreviation and full name}
\item{pop_urb,pop_mun}{The urban and municipal populations (uses 2016
Census data)}
}
}
\usage{
us_cities
}
\description{
This contains a subset of US cities and towns and their populations. This is
for places with a municipal population of greater than 50,000 people
according to 2016 Census data. Obtained from the SimpleMaps website
(https://simplemaps.com).
}
\examples{
# Here is a glimpse at the data
# available in `us_cities`
dplyr::glimpse(us_cities)
}
\keyword{datasets}
