% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vitamind}
\alias{vitamind}
\title{Vitamin D levels across patients}
\format{
A tibble with 533 rows and 6 variables:
\describe{
\item{patient_type}{The type of patient as a factor variable.}
\item{gender}{The reported gender of the patient.}
\item{season}{The season of of the observation as a factor variable.}
\item{region}{The region where the patient resides as a factor variable.}
\item{age}{The age of the patient.}
\item{vitamin_d}{The amount of Vitamin D.}
}
}
\usage{
vitamind
}
\description{
Vitamin D levels across patients
}
\examples{
# Here is a glimpse at the data
# available in `vitamind`
dplyr::glimpse(vitamind)
}
\keyword{datasets}
